+++ lib/grax/grax.ex
@@ -425,4 +425,12 @@ defmodule Grax do
       schema.on_to_rdf(mapping, rdf, opts)
     end
   end
+
+  @spec to_rdf!(Schema.t(), opts :: keyword()) :: Graph.t()
+  def to_rdf!(%schema{} = mapping, opts \\ []) do
+    case to_rdf(mapping, opts) do
+      {:ok, graph} -> graph
+      {:error, error} -> raise error
+    end
+  end
 end